Added payment service for create order saga sample

+++ samples/create-order-saga/ZapMicro.TransactionalOutbox.Samples.CreateOrderSaga.OrderService/Entities/Adjustment.cs
@@ -1,10 +1,11 @@
 using System;
+using ZapMicro.TransactionalOutbox.Samples.CreateOrderSaga.Shared.Entities;
 
 namespace ZapMicro.TransactionalOutbox.Samples.CreateOrderSaga.OrderService.Entities
 {
     public class Adjustment: EntityBase<Guid>
     {
         public string OfferId { get; set; }
-        public decimal Total { get; set; }
+        public double Total { get; set; }
     }
 }
