+++ samples/create-order-saga/ZapMicro.TransactionalOutbox.Samples.CreateOrderSaga.OrderService/Entities/Order.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using ZapMicro.TransactionalOutbox.Samples.CreateOrderSaga.Shared.Entities;
 
 namespace ZapMicro.TransactionalOutbox.Samples.CreateOrderSaga.OrderService.Entities
 {
@@ -9,7 +10,7 @@ namespace ZapMicro.TransactionalOutbox.Samples.CreateOrderSaga.OrderService.Enti
         public OrderStatus Status { get; set; }
         public IList<OrderLine> Lines { get; set; }
         public IList<Adjustment> Adjustments { get; set; }
-        public decimal Total => Lines?.Sum(x => x.FinalTotal) ?? 0;
-        public decimal FinalTotal => Total + Adjustments?.Sum(x => x.Total) ?? 0;
+        public double Total => Lines?.Sum(x => x.FinalTotal) ?? 0;
+        public double FinalTotal => Total + Adjustments?.Sum(x => x.Total) ?? 0;
     }
 }
