+++ samples/create-order-saga/ZapMicro.TransactionalOutbox.Samples.CreateOrderSaga.OrderService/Entities/OrderLine.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using ZapMicro.TransactionalOutbox.Samples.CreateOrderSaga.Shared.Entities;
 
 namespace ZapMicro.TransactionalOutbox.Samples.CreateOrderSaga.OrderService.Entities
 {
@@ -8,10 +9,10 @@ namespace ZapMicro.TransactionalOutbox.Samples.CreateOrderSaga.OrderService.Enti
     {
         public string ProductId { get; set; }
         public uint ProductQuantity { get; set; }
-        public decimal ProductPrice { get; set; }
+        public double ProductPrice { get; set; }
         public IList<Adjustment> Adjustments { get; set; }
-        public decimal Total => ProductQuantity * ProductPrice;
-        public decimal FinalTotal => Total + Adjustments?.Sum(x => x.Total) ?? 0;
+        public double Total => ProductQuantity * ProductPrice;
+        public double FinalTotal => Total + Adjustments?.Sum(x => x.Total) ?? 0;
 
     }
 }
