+++ samples/create-order-saga/ZapMicro.TransactionalOutbox.Samples.CreateOrderSaga.Shared/Events/OnOrderCreated.cs
@@ -5,6 +5,6 @@ namespace ZapMicro.TransactionalOutbox.Samples.CreateOrderSaga.Shared.Events
     public class OnOrderCreated: EventBase
     {
         public Guid OrderId { get; set; }
-        public decimal OrderGrandTotal { get; set; }
+        public double OrderGrandTotal { get; set; }
     }
 }
